agent,billing: move events to pkg/agent/billing
